Change FromValue so it no longer has String, etc. instances. Rename functions.
We now rely on the Text, Inlines, and Blocks instances to parse the relevant
constructors of a MetaValue. Functions parseString, etc., are provided should
such parsers be required.

We also renamed functions so that the from* family (fromValue and so on) now
all have types such as FromValue a => MetaValue -> Result a. The fromObject
function, notably, had to be renamed objectValue so that the fromObject function
could now be in conformity with this convention.