Change error type, improve and reorder documentation.
The error type now has a Maybe MetaErrorMessage in it, and the MetaSomeError has
been removed. MonadFail instances were added to various monads.

The MetaErrorMessage allows us to keep the Expectation of an error while still
having custom messages. We did this by moving the MetaSomeError branch of
MetaError into MetaErrorMessage, which is Either an error message or a
representation of a received value. Custom errors are preferred over received
value messages.

Many parts of the error message that formerly had String in them now have
NonEmptyString or Maybe NonEmptyString, which is a little clearer. This allows
us to filter out empty strings from Expectation, and to filter out empty
components of errors. This affects the throw* family of error throwing functions
as well. throwExpectGot now takes a [String] list of expectations and a String
of what value was received, and filters out empty strings. throwExpectMessage is
new; it takes a [String] of expectations and a String error message and throws a
MetaExpectGotError with a MetaErrorMessage inside it.

While the empty and mzero of Alternative and MonadPlus are still the very
uninformative MetaNullError, we have a MonadFail instance that lets us throw a
custom error message with fail and still preserve other Expectations if
possible.

We also improved simpleErrorShow, partially because we could take advantage of
this new error structure and partially to fix some inelegance in implementation
(we don't display empty components of the error, for instance).

The simple test suite we had had to be modified, since one of our error tests
was affected by the removal of SomeError.

Other than documenting the new features of errors, we improved existing
documentation; in particular, we now clarify that the input of a parser is never
advanced or consumed by parsers and that functions like object and .! are
responsible for advancing deeper into the input. More work might have to be done
to explain this better. A lot of markup had to be added (many things had no
backticks around them).