Change behaviour symbol, symbolfrom, and simpleErrorShow.
Now, we enclose the symbols in double quotation marks when we add them to the
Expectation in thrown errors. This does mean that the expected value shown in
errors won't quite be what is required in the YAML (since the symbollike parser
is laxer than parseString), but it will at least give /a/ valid representation
of the symbol.

We also no longer print the colon after the "when parsing " message in
simpleErrorShow. The field parsers don't yet add a colon after the "field" and
before the "key" in their error message reporting, but they might in the future.